refactor: split tui runtime connection state
